x11/x11proto.h, x11/xutil.c: Use correct parameter type "uint16" for    addsmooth().

+++ trunk/client/x11/xutil.c
@@ -1518,6 +1518,6 @@ void reset_map_data()
 
 
 /* x11 client doesn't do smoothing, so don't need to do anything with this */
-void addsmooth(uint16 face, uint smooth_face)
+void addsmooth(uint16 face, uint16 smooth_face)
 {
 }
